+++ src/Model/Payment/Braintree/Base.php
@@ -32,6 +32,11 @@ use Shopgate\Import\Model\Payment\AbstractPayment;
 class Base extends AbstractPayment
 {
     /**
+     * Map payment method on quote
+     */
+    protected const IS_OFFLINE = false;
+
+    /**
      * The config path for the selected action method
      */
     private const XML_CONFIG_PAYMENT_ACTION = 'payment/%s/payment_action';