+++ src/Test/Integration/Model/Payment/FactoryTest.php
@@ -29,6 +29,7 @@ use Magento\TestFramework\Helper\Bootstrap;
 use Magento\TestFramework\ObjectManager;
 use PHPUnit\Framework\TestCase;
 use Shopgate\Import\Model\Payment\Factory as PaymentFactory;
+use ShopgateLibraryException;
 
 /**
  * @magentoDbIsolation enabled
@@ -53,12 +54,16 @@ class FactoryTest extends TestCase
      * @param string $methodCode
      * @param string $expectedPaymentMethod
      *
-     * @dataProvider paymentMethodProvider
+     * @dataProvider         paymentMethodProvider
      *
      * @throws LocalizedException
+     * @throws ShopgateLibraryException
      *
      * @magentoConfigFixture current_store payment/braintree/active 1
      * @magentoConfigFixture current_store payment/braintree_paypal/active 1
+     * @magentoConfigFixture current_store payment/cashondelivery/active 1
+     * @magentoConfigFixture current_store payment/checkmo/active 1
+     * @magentoConfigFixture current_store payment/banktransfer/active 1
      */
     public function testPaymentMethodMapping($methodCode, $expectedPaymentMethod): void
     {
@@ -75,9 +80,12 @@ class FactoryTest extends TestCase
     public function paymentMethodProvider(): array
     {
         return [
-            'return shopgate as default payment method' => ['PREPAY', 'shopgate'],
+            'return shopgate as default payment method' => ['SOMETHING', 'shopgate'],
+            'Bank Transfer'                             => ['PREPAY', 'banktransfer'],
+            'Cash on delivery'                          => ['COD', 'cashondelivery'],
+            'Invoice'                                   => ['INVOICE', 'checkmo'],
             'Braintree Credit Card'                     => ['BRAINTR_CC', 'braintree'],
-            'Braintree PayPal'                          => ['BRAINTR_PP', 'braintree_paypal']
+            'Braintree PayPal'                          => ['BRAINTR_PP', 'braintree_paypal'],
         ];
     }
 }